+++ pkgs/cronet_http/CHANGELOG.md
@@ -1,3 +1,16 @@
+## 1.2.0
+
+* Support the Cronet embedding dependency with `--dart-define=cronetHttpNoPlay=true`.
+* Fix a bug in the documentation where `isOwned` is used rather than
+  `closeEngine`.
+* Upgrade `package:jni` to 0.7.3 to fix a SIGSEGV caused by a null
+  pointer dereference.
+
+## 1.1.1
+
+* Make it possible to construct `CronetClient` with custom a `CronetEngine`
+  while still allowing `CronetClient` to close the `CronetEngine`.
+
 ## 1.1.0
 
 * Use `package:http_image_provider` in the example application.